add protect condition for edit & delete video

+++ controllers/videoController.js
@@ -60,7 +60,11 @@ export const editVideo = async (req, res) => {
     } = req;
     try {
       const video = await Video.findById(id);
-      res.render("editVideo", { pageTitle: "Edit Video", video });
+      if (!req.user || video.creator !== req.user.id) {
+        throw Error("Unauthenticated Access");
+      } else {
+        res.render("editVideo", { pageTitle: "Edit Video", video });
+      }
     } catch (error) {
       res.redirect(routes.home);
     }
@@ -82,7 +86,12 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (!req.user || video.creator !== req.user.id) {
+      throw Error("Unauthenticated Access");
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
